chore: remove type
